+++ examples/basic/basic.go
@@ -3,10 +3,10 @@ package main
 import "github.com/renatopp/go-cli"
 
 func main() {
-  cli.Name("hello")
-  cli.Description("Prints a classical message.")
-  cli.AutoHelp(true)
-  cli.Parse()
+	cli.Name("hello")
+	cli.Description("Prints a classical message.")
+	cli.AutoHelp(true)
+	cli.Parse()
 
-  println("Hello, World!")
+	println("Hello, World!")
 }