+++ internal/app.go
@@ -70,22 +70,11 @@ func (a *App) Println(format string, v ...any) { a.Stdout(format+"\n", v...) }
 func (a *App) Error(format string, v ...any)   { a.Stderr(format, v...) }
 func (a *App) Errorln(format string, v ...any) { a.Stderr(format+"\n", v...) }
 
-func (a *App) Shell(dir string, name string, args ...string) *exec.Cmd {
+func (a *App) Shell(dir string, name string, args ...string) (string, error) {
 	cmd := exec.Command(name, args...)
-	cmd.Stdout = &StdWriter{fn: func(format string, v ...any) {
-		a.Stdout(format, v...)
-	}}
-	cmd.Stderr = &StdWriter{fn: func(format string, v ...any) {
-		a.Stderr(format, v...)
-	}}
 	cmd.Stdin = os.Stdin
-	return cmd
-}
-
-func (a *App) Exec(dir string, name string, args ...string) error {
-	cmd := a.Shell(dir, name, args...)
-	cmd.Dir = dir
-	return cmd.Run()
+	output, err := cmd.CombinedOutput()
+	return string(output), err
 }
 
 // Parse is called for every command in the path.